Compressor: Add ability to pass uris from a ContentProvider as sourcePath
The current way of retrieving the actual path of the source file is not
reliable across all the different types of Uri that can be used in Android.
The changes here add the option to pass the application context an use the
ContentResolver to obtain a FileDescriptor that can be used as media source.

An example of a an intent Picker that generates such uris is also added to
the demo app.